# refactor(perf): skip first mergeClassList reverse() call