This PR introduces performance optimizations to the `mergeClassList` function

We can eliminate the need for the first array.prototype.reverse() call in the function by concatenating the classnames with `unshift` instead which pushes elements at the start of the array instead of at the end. 

It leads to some performance gains particularly for smaller array lengths (< 5000) which shouldn't be common here, with diminishing returns after that

I also replaced some loops with marginally faster alternatives

my (primitive) [benchmark](https://github.com/rortan134/tailwind-merge/blob/benchmark/tests/benchmark.bench.ts) results are optimistic

before 23.744 ops/sec  0.042 ms ±  15.14 %  (43 runs sampled)
after  56.035 ops/sec  0.018 ms ±  42.14 %  (59 runs sampled)