@XantreDev 2024-07-23
Nice work. Btw, I think it's possible to boost performance further if remove all other unnecessary array allocations. When I've seen code of mergeClassList I was really appalled with amount of overhead