@XantreDev 2024-07-25
> Hey @rortan134! 👋
> 
> Thanks for your patience, just came back from a vacation. Cool that you took the time to make tailwind-merge faster! I'm definitely up to adding this change if it makes things that much faster.
> 
> [I tried to do a similar perf optimization in the past](https://github.com/dcastil/tailwind-merge/pull/20) but it turned out not to make a difference, probably due to the usually small arrays that are handled there (<100 elements). But I also didn't use a benchmark suite but measured the performance in a real app that used tailwind-merge a lot.
> 
> A backwards running for loop with simple string concatenation like I did back then should theoretically be the fastest way to do the work in `mergeClassList` since there are no array allocations anymore other than splitting the original string. That also eliminates the `result.unshift()` call which can also become quite slow. I'm not 100% sure how up to date this info is with JS engines doing all sorts of tricks nowadays, but in a simple implementation of `unshift()` all existing array elements need to be moved in order to add one at the start, wasting a lot of CPU cyles. More on that in https://stackoverflow.com/questions/44031591/performance-of-array-push-vs-array-unshift.
> 
> Could you test how the code performs when you switch to the backwards running for loop with string concatenation as described in [#20 (comment)](https://github.com/dcastil/tailwind-merge/pull/20#issue-978463742)? Is that faster or slower than your approach?

Yep, seems to be unshift is O(n)