@XantreDev 2024-07-26
There is no much performance difference. I think most of the difference on side of memory usage, don't know how to bench it yet
![image](https://github.com/user-attachments/assets/ab3324d8-6e26-4fac-9eb4-9f11b24ca15b)
Bench is [here](https://github.com/XantreDev/tailwind-merge/tree/bench)