@dcastil 2024-07-28
> `mergeClassList` will also be poorly JIT-ed, because it's not possible to inline `parseClassName`, `getClassGroupId`, `getConflictingClassGroupIds`, since it's passed as parameter

@XantreDev What do you mean with "poorly JIT-ed"? I know what JIT is but don't know much about about how JS engines compile down JS code.

> There is no much performance difference. […]

Thanks for your benchmarks! Your relative results seem to align with my own results in https://github.com/dcastil/tailwind-merge/pull/20#issue-978463742. That means the string concatenation approach is the best one, although the difference is not too big.

Is there a reason why you used an array instead of a set for [`classGroupsInConflict`](https://github.com/XantreDev/tailwind-merge/blob/ef59466ad6053aaa2803d4aafb4844676d161b71/src/lib/merge-classlist.ts#L110)? I think a set should perform better in both time and space there because of the O(1) lookup and duplicates removal.