@dcastil 2024-07-28
@rortan134 would you be willing to adopt the string concatenation approach from https://github.com/XantreDev/tailwind-merge/blob/ef59466ad6053aaa2803d4aafb4844676d161b71/src/lib/merge-classlist.ts#L100-L179? Looks like that is the best approach.