@XantreDev 2024-07-28
> Is there a reason why you used an array instead of a set for [`classGroupsInConflict`](https://github.com/XantreDev/tailwind-merge/blob/ef59466ad6053aaa2803d4aafb4844676d161b71/src/lib/merge-classlist.ts#L110)? I think a set should perform better in both time and space there because of the O(1) lookup and duplicates removal.

It depends from amount of data we store in set. Array will perform better with relatively small number up elements (up to hundreds of elements). Probably you're right if we are talking about strings, but there was no difference in benchmarks of this code