@XantreDev 2024-07-28
> > `mergeClassList` will also be poorly JIT-ed, because it's not possible to inline `parseClassName`, `getClassGroupId`, `getConflictingClassGroupIds`, since it's passed as parameter
> 
> @XantreDev What do you mean with "poorly JIT-ed"? I know what JIT is but don't know much about about how JS engines compile down JS code.

JIT optimizer can optimize some levels of abstraction if it will infer types correctly, but here is kind of hard to infer types, because a lot of code in polymorphic, there are also a lot of memory allocations is going on. JIT also cannot inline some functions, because it's passed as a parameter (I've experemented with that a bit and didn't found any performance benefit from using additional closure)

I will think about how to make this library much faster, because 32k ops/s is not really great. It can use a lot of performance budget when we rendering on mobile with 120 hz display