@dcastil 2024-07-29
Hey @rortan134, I just merged https://github.com/dcastil/tailwind-merge/pull/450 with the improvements discussed in this PR, so I'm closing this one.

I'll include you in the release notes to acknowledge your contribution as well. Thanks you for your PR!